SD-JWT: Extract creating SD JWT into separate file
