+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/CredentialToBeIssued.kt
@@ -33,5 +33,6 @@ sealed class CredentialToBeIssued {
  * Represents a claim that shall be issued to the holder,
  * i.e. serialized into the appropriate credential format.
  * To issue nested structures in SD-JWT, pass a Collection of [ClaimToBeIssued] in [value].
+ * For each claim, one can select if the claim shall be selectively disclosable, or otherwise included plain.
  */
 data class ClaimToBeIssued(val name: String, val value: Any, val selectivelyDisclosable: Boolean = true)
